# perf/merkle inplace