--- a/crypto/Cargo.toml
+++ b/crypto/Cargo.toml
@@ -15,5 +15,17 @@ thiserror = "1.0.38"
 [dependencies.rand]
 version = "0.8"
 
+[dev-dependencies]
+criterion = "0.4"
+iai-callgrind.workspace = true
+
 [features]
 test_fiat_shamir = []
+
+[[bench]]
+name = "criterion_merkle"
+harness = false
+
+[[bench]]
+name = "iai_merkle"
+harness = false