--- a/crypto/benches/criterion_merkle.rs
+++ b/crypto/benches/criterion_merkle.rs
@@ -19,10 +19,8 @@ fn merkle_tree_benchmarks(c: &mut Criterion) {
 
     // NOTE: the values to hash don't really matter, so let's go with the easy ones.
     let unhashed_leaves: Vec<_> = core::iter::successors(Some(FE::zero()), |s| Some(s + FE::one()))
-        // `(1 << 10) + 1` exploits worst cases in terms of rounding up to powers of 2.
-        // Typical size should really be about `1 << 20`, but it currently takes too long
-        // to make a single iteration with such numbers.
-        .take((1 << 10) + 1)
+        // `(1 << 20) + 1` exploits worst cases in terms of rounding up to powers of 2.
+        .take((1 << 20) + 1)
         .collect();
 
     group.bench_with_input("build", unhashed_leaves.as_slice(), |bench, unhashed_leaves| {