- perf: add benchmark for merkle tree construction
- perf: lower number of elements for merkle bench
- perf: remove allocations in recursive merkle build step
- perf: bump sample size again, as now it can finish

## Type of change

- [x] Optimization

## Checklist
- [x] This change is an Optimization
  - [x] Benchmarks added

## Results

Before: couldn't finish the warmup with 1M elements.
After:
```
Benchmarking Merkle Tree/build: Warming up for 3.0000 s
Merkle Tree/build       time:   [1.0337 s 1.0367 s 1.0381 s]
```